Actual bean project with spring
Stepped back and did a very very very basic project utilizing beans and spring.

This one is fully functional, doesn't have a web dependency (which is where most of my problems stemmed from). It does 2 things.

It prints "Hello" followed by a value taken from the bean create, and then the bean is set to a value in the main, and says "Hello" there

After that, I created a new bean, and gave it 2 values (X,Y). The values are calculated through the bean and displayed in console via addition, subtraction, multi, and division.