HashSet Added
Hashset added

That's it, not gonna bother with a while loop or anything, just doing this as PoC. Last update for the night, I was gonna add another bean but I'm not sure what else to do.

I say this as I will probably do some more work tonight like say, making a bean that will hold user inputs and working across all the beans.